requirements.txt
# Python dependencies for publication updates
scholarly>=1.5.0
PyYAML>=6.0
beautifulsoup4>=4.11.0
requests>=2.28.0